github_events_scraper/pyproject.toml
[tool.poetry]
name = "github-events-scraper"
version = "0.1.0"
description = "Scrapes Github events for each configured repository to be aggregated later."
authors = ["Richard Volcko <rikvolcko@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
toml = "^0.10.2"
requests = "^2.32.3"
pydantic = "^2.11.2"
sqlalchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
